backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.2
black==23.11.0
bleach==6.2.0
celery==5.4.0
celery-types==0.23.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.2
cryptography==46.0.2
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-ratelimit==4.1.0
django-redis==5.4.0
django-stubs==4.2.6
django-stubs-ext==5.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.14.2
drf-spectacular==0.28.0
factory-boy==3.3.0
Faker==37.5.3
flake8==6.1.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.12.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
mccabe==0.7.0
mypy==1.7.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.23
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
transliterate==1.10.2
types-pytz==2025.2.0.20250809
types-PyYAML==6.0.12.20250809
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.6.0
